1.23

+++ src/project01/CustomerView.java
@@ -5,7 +5,34 @@ public class CustomerView {
     private CustomerList customers = new CustomerList(10);
 
     public void enterMainMenu() {
-
+        while (true){
+            System.out.println("-----------------客户信息管理软件-----------------" +
+                    "\n" +
+                    "                            1 添 加 客 户\n" +
+                    "                            2 修 改 客 户\n" +
+                    "                            3 删 除 客 户\n" +
+                    "                            4 客 户 列 表\n" +
+                    "                            5 退           出\n" +
+                    "\n" +
+                    "                            请选择(1-5)：_");
+            int number=CMUtility.readInt(5);
+            switch (number){
+                case 1:
+                    addNewCustomer();
+                    break;
+                case 2:
+                    modifyCustomer();
+                    break;
+                case 3:
+                    deleteCustomer();
+                    break;
+                case 4:
+                    listAllCustomers();
+                    break;
+                default:
+                    break;
+            }
+        }
     }
 
     private void addNewCustomer() {
@@ -21,7 +48,7 @@ public class CustomerView {
         System.out.print("邮箱：");
         String mail = CMUtility.readString(15, "无邮箱");
         System.out.println("---------------------添加完成---------------------");
-        Customer customer = new Customer(customers.getAllCustomers().length, name, gender, age, phone_number, mail);
+        Customer customer = new Customer(name, gender, age, phone_number, mail);
         customers.addCustomer(customer);
     }
 
@@ -41,13 +68,28 @@ public class CustomerView {
         System.out.print("邮箱：(" + customer.getEmail() + ")：");
         String mail = CMUtility.readString(12, customer.getEmail());
         System.out.println("---------------------修改完成---------------------");
-        customers.replaceCustomer(id, new Customer(id, name, gender, age, phone, mail));
+        customers.replaceCustomer(id, new Customer(name, gender, age, phone, mail));
     }
 
     private void deleteCustomer(){
         System.out.println("---------------------删除客户--------------------- ");
         System.out.println("请选择待删除客户编号(-1退出)：");
         int index=CMUtility.readInt(-1);
-        customers.deleteCustomer(index);
+        System.out.println("确认是否删除(Y/N)：");
+        char res=CMUtility.readConfirmSelection();
+        if (res=='Y'||res=='y'){
+            customers.deleteCustomer(index);
+        }
+        System.out.println("---------------------删除完成---------------------\n");
+    }
+    private void listAllCustomers(){
+        System.out.println("---------------------------客户列表---------------------------");
+        System.out.println("编号\t姓名\t性别\t年龄\t电话\t邮箱");
+        Customer[] list=customers.getAllCustomers();
+        for (Customer customer : list) {
+            System.out.println(customer.getId()+"\t\t"+customer.getName()+"\t\t"+customer.getGender()+"\t\t"+customer.getAge()
+            +"\t\t"+customer.getPhone()+"\t\t"+customer.getEmail());
+        }
+        System.out.println("-------------------------客户列表完成-------------------------\n");
     }
 }
